Eliminar archivos al actualizar o eliminar productos

+++ controllers/admin.js
@@ -1,7 +1,7 @@
 const { default: mongoose } = require('mongoose');
 const Producto = require('../models/producto');
 const { validationResult } = require('express-validator');
-
+const file = require('../utils/file')
 
 exports.getCrearProducto = (req, res) => {
     res.render('admin/editar-producto', {
@@ -139,6 +139,7 @@ exports.postEditarProducto = (req, res, next) => {
             producto.precio = precio;
             producto.descripcion = descripcion;
             if (imagen) {
+                file.deleteFile(producto.urlImagen);
                 producto.urlImagen = imagen.path;
             }
             return producto.save();
@@ -175,7 +176,15 @@ exports.getProductos = (req, res, next) => {
 
 exports.postEliminarProducto = (req, res, next) => {
     const idProducto = req.body.idProducto;
-    Producto.deleteOne({ _id: idProducto, idUsuario: req.usuario._id })
+
+    Producto.findById(idProducto)
+        .then(producto => {
+            if(!producto) {
+                return next(new Error('Producto no se ha encontrado'));
+            }
+            file.deleteFile(producto.urlImagen);
+            return Producto.deleteOne({ _id: idProducto, idUsuario: req.usuario._id });
+        })
         .then(result => {
             res.redirect('/admin/productos');
         })
